+++ IngSoft.UI/DTOs/BitacoraGridDto.cs
@@ -4,7 +4,6 @@ namespace IngSoft.UI.DTOs
 {
     internal class BitacoraGridDto
     {
-        public Guid Id { get; set; }
         public DateTime Fecha { get; set; }
         public string Descripcion { get; set; }
         public string Origen { get; set; }
