--- IngSoft.UI/Form1.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Windows.Forms;
-
-namespace IngSoft.UI
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-
-        private void Form1_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btnBitacora_Click(object sender, EventArgs e)
-        {
-            var frmBitacora = new FrmBitacora();
-            frmBitacora.ShowDialog();
-        }
-
-        private void btnUsuario_Click(object sender, EventArgs e)
-        {
-            var frmUsuario = new FrmUsuario();
-            frmUsuario.ShowDialog();
-        }
-    }
-}
