+++ IngSoft.UI/FrmPrincipal.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Windows.Forms;
+
+namespace IngSoft.UI
+{
+    public partial class FrmPrincipal : Form
+    {
+        public FrmPrincipal()
+        {
+            InitializeComponent();
+        }
+
+        private void Form1_Load(object sender, EventArgs e)
+        {
+
+        }
+        private void usuariosToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            var frmUsuario = new FrmUsuario();
+            frmUsuario.ShowDialog();
+        }
+
+        private void bitacoraToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            var frmBitacora = new FrmBitacora();
+            frmBitacora.ShowDialog();
+        }
+    }
+}
