+++ IngSoft.UI/Program.cs
@@ -13,7 +13,7 @@ namespace IngSoft.UI
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new FrmPrincipal());
         }
     }
 }
